:recycle: Create component patientLine fix #71
